Apply Result type for flow control

+++ Cornershop.Service.Domain/Services/ProductService.cs
@@ -11,7 +11,7 @@ namespace Cornershop.Service.Domain.Services;
 
 public class ProductService(IDbContextFactory<CornershopDbContext> dbContextFactory) : IProductService
 {
-    public async Task<ProductDTO?> GetById(string id, bool isHiddenIncluded = false)
+    public async Task<Result<ProductDTO?, string?>> GetById(string id, bool isHiddenIncluded = false)
     {
         using var dbContext = await dbContextFactory.CreateDbContextAsync();
         if (isHiddenIncluded)
@@ -21,7 +21,8 @@ public class ProductService(IDbContextFactory<CornershopDbContext> dbContextFact
                 .Include(p => p.Publisher)
                 .Include(p => p.Subcategory).ThenInclude(s => s.Category)
                 .Include(p => p.ProductImages)
-                .FirstOrDefaultAsync() ?? throw new Exception(); //TO BE FIXED
+                .FirstOrDefaultAsync();
+            if (product == null) return Constants.ERR_PRODUCT_NOT_FOUND;
             return product.Map();
         }
         else
@@ -31,7 +32,8 @@ public class ProductService(IDbContextFactory<CornershopDbContext> dbContextFact
                 .Include(p => p.Publisher)
                 .Include(p => p.Subcategory).ThenInclude(s => s.Category)
                 .Include(p => p.ProductImages)
-                .FirstOrDefaultAsync() ?? throw new Exception(); //TO BE FIXED
+                .FirstOrDefaultAsync();
+            if (product == null) return Constants.ERR_PRODUCT_NOT_FOUND;
             return product.Map();
         }
     }
@@ -81,13 +83,16 @@ public class ProductService(IDbContextFactory<CornershopDbContext> dbContextFact
         }
     }
 
-    public async Task<ProductDTO?> Add(ProductDTO productDTO)
+    public async Task<Result<ProductDTO?, string?>> Add(ProductDTO productDTO)
     {
         using var dbContext = await dbContextFactory.CreateDbContextAsync();
 
-        var subcategory = await dbContext.Subcategories.FirstOrDefaultAsync(s => s.Id == productDTO.SubcategoryId) ?? throw new Exception(); //TO BE FIXED
-        var author = await dbContext.Authors.FirstOrDefaultAsync(a => a.Id == productDTO.AuthorId) ?? throw new Exception();
-        var publisher = await dbContext.Publishers.FirstOrDefaultAsync(p => p.Id == productDTO.PublisherId) ?? throw new Exception(); //TO BE FIXED
+        var subcategory = await dbContext.Subcategories.FirstOrDefaultAsync(s => s.Id == productDTO.SubcategoryId);
+        if (subcategory == null) return Constants.ERR_SUBCATEGORY_NOT_FOUND;
+        var author = await dbContext.Authors.FirstOrDefaultAsync(a => a.Id == productDTO.AuthorId);
+        if (author == null) return Constants.ERR_AUTHOR_NOT_FOUND;
+        var publisher = await dbContext.Publishers.FirstOrDefaultAsync(p => p.Id == productDTO.PublisherId);
+        if (publisher == null) return Constants.ERR_AUTHOR_NOT_FOUND;
 
         var product = new Product
         {
@@ -140,7 +145,7 @@ public class ProductService(IDbContextFactory<CornershopDbContext> dbContextFact
         return product.Map();
     }
 
-    public async Task<ProductDTO?> Update(ProductDTO productDTO)
+    public async Task<Result<ProductDTO?, string?>> Update(ProductDTO productDTO)
     {
         using var dbContext = await dbContextFactory.CreateDbContextAsync();
         var product = await dbContext.Products
@@ -148,10 +153,14 @@ public class ProductService(IDbContextFactory<CornershopDbContext> dbContextFact
             .Include(p => p.Publisher)
             .Include(p => p.Subcategory)
             .Include(p => p.ProductImages)
-            .FirstOrDefaultAsync(p => p.Id == productDTO.Id) ?? throw new Exception(); //TO BE FIXED
+            .FirstOrDefaultAsync(p => p.Id == productDTO.Id);
+        if (product == null) return Constants.ERR_PRODUCT_NOT_FOUND;
         var subcategory = await dbContext.Subcategories.FirstOrDefaultAsync(p => p.Id == productDTO.SubcategoryId);
+        if (subcategory == null) return Constants.ERR_SUBCATEGORY_NOT_FOUND;
         var author = await dbContext.Authors.FirstOrDefaultAsync(p => p.Id == productDTO.AuthorId);
+        if (author == null) return Constants.ERR_AUTHOR_NOT_FOUND;
         var publisher = await dbContext.Publishers.FirstOrDefaultAsync(p => p.Id == productDTO.PublisherId);
+        if (publisher == null) return Constants.ERR_PUBLISHER_NOT_FOUND;
 
         product.Name = productDTO.Name ?? product.Name;
         product.Description = productDTO.Description ?? product.Description;
@@ -197,10 +206,11 @@ public class ProductService(IDbContextFactory<CornershopDbContext> dbContextFact
         return product.Map();
     }
 
-    public async Task<bool> Remove(string id)
+    public async Task<Result<bool, string?>> Remove(string id)
     {
         using var dbContext = await dbContextFactory.CreateDbContextAsync();
-        var product = await dbContext.Products.FirstOrDefaultAsync(p => p.Id == id) ?? throw new Exception(); //TO BE FIXED
+        var product = await dbContext.Products.FirstOrDefaultAsync(p => p.Id == id);
+        if (product == null) return Constants.ERR_PRODUCT_NOT_FOUND;
         dbContext.Products.Remove(product);
         await dbContext.SaveChangesAsync();
         return true;
